feat(trace-utils): widen trace id to 128 bits
# Motivation

When we serialize we store the uppper 64 bits  of the trace id in a meta tag. But at runtime, we don't want to split it as the trace id still needs to be readable.

+++ datadog-trace-utils/src/msgpack_encoder/v04/span.rs
@@ -207,7 +207,7 @@ pub fn encode_span<W: RmpWrite, T: SpanText>(
     write_str(writer, span.resource.borrow())?;
 
     write_str(writer, "trace_id")?;
-    write_u64(writer, span.trace_id)?;
+    write_u64(writer, span.trace_id as u64)?;
 
     write_str(writer, "span_id")?;
     write_u64(writer, span.span_id)?;
